% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_snps.R
\name{annotate_snps}
\alias{annotate_snps}
\title{Annotate SNPs from genetic fine-mapping or GWAS index variants}
\usage{
annotate_snps(
  snp_file,
  tissue_path_file,
  tissue_annotation_file,
  genomic_path_file,
  genomic_annotation_file
)
}
\arguments{
\item{snp_file}{Path to the input SNP file}

\item{tissue_path_file}{Path to the input file of paths to tissue annotation bed files}

\item{tissue_annotation_file}{Path to the input file of tissue annotation names and weights}

\item{genomic_path_file}{Path to the input file of paths to genomic annotation bed files}

\item{genomic_annotation_file}{Path to the input file of genomic annotation names and weights}
}
\description{
This function annotates SNPs to genomic annotations (e.g. coding sequence) and
tissue-level annotations (e.g. chromatin states). Each SNP is mapped against each
set of annotations contained within BED files listed within the genomic annotation file and
tissue annotation file. The annotation names listed in the annotation files must be consistent
with the annotation names listed within the BED files.
}
