% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_tissue_vectors.R
\name{calculate_tissue_vectors}
\alias{calculate_tissue_vectors}
\title{Calculate a tissue score vector for each SNP within a genetic credible set or for each index SNP}
\usage{
calculate_tissue_vectors(
  snp.annotated.df,
  tissue_annotation_file,
  genomic_annotation_file,
  ess.annot = NULL,
  ess.file = NULL
)
}
\arguments{
\item{snp.annotated.df}{Dataframe of annotated SNP info, this is output from annotate_snps function}

\item{tissue_annotation_file}{Path to the input file of tissue annotation names and weights}

\item{genomic_annotation_file}{Path to the input file of genomic annotation names and weights]}

\item{ess.annot}{Optional: Name of genomic annotation requiring specificity scores (e.g. "coding")}

\item{ess.file}{Optional: Path to the specificity score file for ess.annot, (e.g. expression specificity scores for coding annotations)}
}
\description{
This function takes the SNP annotations outputted from the annotate_snps function and the
provided annotation weigths to yield a set of tissue scores that indicate how much of the
fine-mapping/association signal can be apportioned to each tissue/cell type
}
