% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_toa_scores.R
\name{calculate_toa_scores}
\alias{calculate_toa_scores}
\title{Calculate a tissue-of-action (TOA) scores for each genetic signal}
\usage{
calculate_toa_scores(snp.tissvec.df)
}
\arguments{
\item{snp.tissvec.df}{Dataframe of tissue scores for each SNP (output from calculate_tissue_vectors function)}
}
\description{
This function takes the tissue scores for each individual SNP outputted from
calculate_tissue_vectors function and calculates a TOA score for each tissue/cell type for each
genetic signal. A signal would correspond to a credible set from genetic fine-mapping or simply
an index SNP from GWAS
}
