% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissue_classifier.R
\name{tissue_classifer}
\alias{tissue_classifer}
\title{Classify genetic signals using a simple rule-based classifier}
\usage{
tissue_classifer(toa.df, tissue_threshold = 0.2, shared_threshold = 0.1)
}
\arguments{
\item{toa.df}{Dataframe of tissue scores for each SNP (output from calculate_tissue_vectors function)}

\item{tissue_threshold}{Dataframe of tissue scores for each SNP (output from calculate_tissue_vectors function)}

\item{shared_threshold}{Dataframe of tissue scores for each SNP (output from calculate_tissue_vectors function)}
}
\description{
This function takes the calculated toa scores for each genetic signal and applies a simple
rule-based classifier to assign each signal to a tissue/cell-type. The signal is assigned to the
tissue/cell-type that has the highest TOA score and exceeds a user-specified score threshold (default 0.2).
The user can also specify a threshold for determining shared signals. If two or more tissues yield TOA scores
that fall within this specified range of each other, the signal is designated as a "shared" signal. If
designated as shared, the responsible tissues will be indicated in the output data frame.
}
